requirements-bot.txt
# Telegram Bot dependencies
aiogram==3.4.1

# Google Cloud - updated for Gemini 2.5 support
google-cloud-aiplatform>=1.50.0
google-cloud-firestore>=2.14.0
google-cloud-secret-manager>=2.18.1

# Utils
aiohttp==3.9.3
Pillow==10.2.0
pydantic-settings>=2.1.0